Improvements to new schemas

+++ sql/V21.03__create_rmas_table.sql
@@ -1,6 +1,6 @@
 create table rmas (
     id serial primary key,
-    order_id bigint not null,
+    order_id int not null,
     reason character varying(255), -- Placeholder
     source character varying(255), -- Placeholder
     foreign key (order_id) references orders(id) on update restrict on delete restrict
