+++ sql/V21.04__create_rma_receipts_table.sql
@@ -1,6 +1,6 @@
 create table rma_receipts (
     id bigint primary key,
-    rma_id bigint not null,
+    rma_id int not null,
     receiver_name character varying(255), -- Placeholder
     inventory_location_id int, -- Placeholder
     created_at timestamp without time zone default (now() at time zone 'utc'),
